rename popularlist replyList
